changed magnetize_func to int, for now later I'll maybe add a proper interface to keep particle attributes light, but be able to input and output strings as name of the functions to use e.g. "langevin" translates to uint8_t of value 0

+++ src/core/Particle.hpp
@@ -240,7 +240,7 @@ struct ParticleProperties {
 
 #ifdef ESPRESSO_MAGNETIZE
   bool is_magnetizable = false;
-  uint8_t magnetize_func = 0;
+  int magnetize_func = 0;
 #endif
 
   template <class Archive> void serialize(Archive &ar, long int /* version */) {
